+++ CHANGELOG.md
@@ -3,16 +3,22 @@
 Todas as mudanças notáveis neste projeto são documentadas neste arquivo.
 O formato é baseado em [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.2.0] - 2025-11-04
+
+### Added
+- Stream e Map para organizar os tipos de naipe em ordem alfabética e atribuir um index para cada constante da classe
+- Pré-lógica para criação de cartas com diferentes naipes (cada naipe tem 13 cartas)
+
 ---
 
-## [0.1.1] - 2025-10-04
+## [0.1.1] - 2025-11-04
 
 ### Removed
 - Pasta .idea adicionada acidentalmente pelo IntelliJ IDEA
 
 ---
 
-## [0.1.0] - 2025-10-04
+## [0.1.0] - 2025-11-04
 
 ### Added
 - Estrutura inicial da mesa de jogo e de classes que compõem o baralho.